onclite: overlay: Disable vendor mismatch warning
* Remove the vendor mismatch message due to custom kernel

For Non Treble devices when build fingerprint is modified, At the initial boot Android throws up the error vendor image does not match with the system.
Setting to false by default, as no one would like to see that message.
Sometimes this annoying message appears depending on the kernel you want to use or change.

Change-Id: I95503c20b57d56fe7d3142cef39c6e8b3b24ccdb
Signed-off-by: rayrzy <vincentrayananda@gmail.com>